+++ instapy/comment_util.py
@@ -13,7 +13,6 @@ from .util import explicit_wait
 from .util import extract_text_from_element
 from .util import web_address_navigator
 from .event import Event
-from .like_util import get_media_edge_comment_string
 from .quota_supervisor import quota_supervisor
 from .xpath import read_xpath
 
@@ -211,7 +210,13 @@ def verify_mandatory_words(mand_words, comments, browser, logger,):
         if post_desc is None and first_comment is None:
             return False, [], "couldn't get post description and comments"
 
-        text = post_desc if post_desc is not None else '' + ' ' + first_comment if first_comment is not None else ''
+        text = (
+            post_desc
+            if post_desc is not None
+            else "" + " " + first_comment
+            if first_comment is not None
+            else ""
+        )
 
         if len(mand_words) > 0:
             if not evaluate_mandatory_words(text, mand_words):
@@ -374,8 +379,8 @@ def get_comments_count(browser, logger):
             ".graphql.shortcode_media.edge_media_preview_comment.count"
         )
 
-        #media_edge_string = get_media_edge_comment_string(media)
-        #comments_count = media[media_edge_string]["count"]
+        # media_edge_string = get_media_edge_comment_string(media)
+        # comments_count = media[media_edge_string]["count"]
 
     except Exception as e:
         try:
@@ -399,18 +404,25 @@ def get_comments_count(browser, logger):
     return comments_count, "Success"
 
 
-def process_comments(comments, clarifai_comments, delimit_commenting, max_comments, min_comments, comments_mandatory_words, user_name, blacklist, browser, logger, logfolder, publish=True):
+def process_comments(
+    comments,
+    clarifai_comments,
+    delimit_commenting,
+    max_comments,
+    min_comments,
+    comments_mandatory_words,
+    user_name,
+    blacklist,
+    browser,
+    logger,
+    logfolder,
+    publish=True,
+):
 
     # comments
     if delimit_commenting:
-        (
-            commenting_approved,
-            disapproval_reason,
-        ) = verify_commenting(
-            browser,
-            max_comments,
-            min_comments,
-            logger,
+        (commenting_approved, disapproval_reason,) = verify_commenting(
+            browser, max_comments, min_comments, logger,
         )
     if not commenting_approved:
         logger.info(disapproval_reason)
@@ -420,12 +432,7 @@ def process_comments(comments, clarifai_comments, delimit_commenting, max_commen
         commenting_approved,
         selected_comments,
         disapproval_reason,
-    ) = verify_mandatory_words(
-        comments_mandatory_words,
-        comments,
-        browser,
-        logger,
-    )
+    ) = verify_mandatory_words(comments_mandatory_words, comments, browser, logger,)
     if not commenting_approved:
         logger.info(disapproval_reason)
         return False
@@ -436,11 +443,6 @@ def process_comments(comments, clarifai_comments, delimit_commenting, max_commen
     # smart commenting
     if comments and publish:
         comment_state, msg = comment_image(
-            browser,
-            user_name,
-            selected_comments,
-            blacklist,
-            logger,
-            logfolder,
+            browser, user_name, selected_comments, blacklist, logger, logfolder,
         )
         return comment_state