+++ tests/test_file_handling.py
@@ -7,18 +7,17 @@ class TestFileHandling(unittest.TestCase):
     """Unit tests for file_handling function"""
 
     # list of expected entries
-    global expected_entries
     expected_entries = ["This is a comment...", "photography", "portrait"]
 
     def test_valid_entries(self):
         """Check file_handling with valid entries"""
         actual_entries = util.file_handling("tests/resources/valid_target_list.txt")
-        self.assertListEqual(actual_entries, expected_entries)
+        self.assertListEqual(actual_entries, TestFileHandling.expected_entries)
 
     def test_invalid_entries(self):
         """Check file_handling with invalid entries"""
         actual_entries = util.file_handling("tests/resources/invalid_target_list.txt")
-        self.assertListEqual(actual_entries, expected_entries)
+        self.assertListEqual(actual_entries, TestFileHandling.expected_entries)
 
     def test_invalid_filepath(self):
         """Check whether it raises an exception when the filepath is invalid"""