+++ CHANGELOG.md
@@ -5,12 +5,17 @@ The **goal** of this file is explaining to the users of our project the notable 
 _The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)_.
 
 
-## [Unreleased]
+## [0.6.9] - 2020-06-12
 
 ### Added
 
 - Additional parameter `browser_executable_path` now available when initializing InstaPy. Use it to run a specific installation of Firefox.
 - A new feature - `target_list()` to parse text files containing target lists of users, hashtags etc.
+
+
+### Changed
+- Remove `view-source` which stops bot from proceeding 
+- Remove instagram status check
 
 
 ## [0.6.8] - 2020-01-28